requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
PyJWT==2.1.0
psycopg2-binary==2.9.1
werkzeug==2.0.3
Flask-Migrate
SQLAlchemy<2.0

